requirements.txt
python-dotenv
requests
gradio
pillow
flask
transformers
torch
openai

seaborn
pandas
matplotlib
numpy
scipy